Update introduction_to_css.md
Completed copyedit on file. NOTE: three things: 1. Line 13 repeats what is already explained in line 5. I suggest removing one. 2. the title in line 9 references CSS3, but  never addresses it in the copy. I left the reference because it might be intentional, but it does need content added to justify it being there. 3. "what you learned" doesn't really describe what was learned, but rather reminders. I added summary content of what was learned, so please review and modify as needed.