Update serviceable_urls.md
Completed copyedit on file. NOTE: I changed the page title substantially because I believe the existing title did not represent the content directly. Also, review my changes to "Get " requests in the last paragraph to ensure I didn't change required formatting to that element of code (don't think I did based on the content in the "Getters and Setters" file, but wanted to be sure).