Update 2-NumPy.ipynb
Completed copyedit on file. NOTE: In most instances where a specific "section" was mentioned, I changed the reference to "module" and added a link to it.  Also, lines 1324-25...it seems a word is missing at the end of the first line. In line 2015, the first sentence needs help. It wasn't clear as is but I don't know if my edit helped.